Update theme to Fork CMS 3.7.x
I capitalized/CamelCased the folder names and some files to be compatible with Fork CMS 3.7.x just like the Triton theme. I also changed each `core/layout/templates` path in the templates to the capitalized version, etc...

I also removed the Fancybox css import line of code in `screen.css` because it was throwing an error in the browser console (the file doesn't exist), and fancybox isn't used in the theme (no js to be found).